+++ include/CString.h
@@ -141,7 +141,7 @@ class CString: public std::string
         {
             CString strTemp;
             strTemp.reserve(size());
-            for(int i = 0; i < size(); ++i)
+            for(size_t i = 0; i < size(); ++i)
                 strTemp += (char)tolower(at(i));
             return strTemp;
         }
